coap-server/requirements.txt
aiocoap
paho-mqtt

# Needed for --pretty-print
cbor2
pygments
cbor-diag
